requirements.txt
python-telegram-bot==20.7
groq==0.4.1
python-dotenv==1.0.1
python-magic==0.4.27
chromadb==0.4.22
docker==7.0.0
rank-bm25==0.2.2
tiktoken==0.5.2
ollama==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1
yt-dlp==2024.12.23
requests==2.31.0
